Moved App.vue to components
