+++ frontend/js/app.js
@@ -1,6 +1,6 @@
 import "./bootstrap";
 import { createApp } from "vue";
-import App from "@/App.vue";
+import App from "@/components/App.vue";
 import router from "@/router/index.js";
 import "@/assets/tailwind.css";
